% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wormypath-package.R
\docType{package}
\name{wormypath-package}
\alias{wormypath}
\alias{wormypath-package}
\title{wormypath: Wormy Paths}
\description{
This will make wormy paths.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JoFrhwld/wormypath}
  \item Report bugs at \url{https://github.com/JoFrhwld/wormypath/issues}
}

}
\author{
\strong{Maintainer}: Josef Fruehwald \email{jofrhwld@gmail.com}

}
\keyword{internal}
